Refactor cli into a service
