+++ container.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { Config } from './config';
+import { Config, Argv, Env } from './config';
+import { Main } from './cli';
 import { ConsoleHost } from './host';
 import { Editor } from './editor';
 import { Executor } from './executor';
@@ -9,9 +10,10 @@ import { Executor } from './executor';
   providers: [
     { provide: 'argv', useValue: process.argv.slice(2) },
     { provide: 'env', useValue: process.env },
+    { provide: 'exitFn', useValue: process.exit },
     {
       provide: Config,
-      useFactory: (argv: typeof process.argv, env: typeof process.env) => {
+      useFactory: (argv: Argv, env: Env) => {
         return Config.load(argv, env);
       },
       inject: ['argv', 'env'],
@@ -22,6 +24,7 @@ import { Executor } from './executor';
     },
     Editor,
     Executor,
+    Main,
   ],
 })
 export class Container {}
