+++ main.ts
@@ -1,6 +1,3 @@
 import { main } from './index';
 
-main({
-  argv: process.argv.slice(2),
-  env: process.env,
-});
+main();
